+++ Tests/RelayerTests/AuthTests/EdDSASignerTests.swift
@@ -1,0 +1,13 @@
+import Foundation
+import XCTest
+import WalletConnectKMS
+@testable import WalletConnectRelay
+
+final class EdDSASignerTests: XCTestCase {
+    var sut: EdDSASigner!
+    let expectedSignature = "0JkxOM-FV21U7Hk-xycargj_qNRaYV2H5HYtE4GzAeVQYiKWj7YySY5AdSqtCgGzX4Gt98XWXn2kSr9rE1qvCA"
+
+    func test() {
+
+    }
+}
