+++ Tests/RelayerTests/AuthTests/JWTTests.swift
@@ -18,13 +18,3 @@ final class JWTTests: XCTestCase {
     }
 
 }
-
-class EdDSASignerMock: JWTSigning {
-    var alg: String = "EdDSA"
-
-    func sign(header: String, claims: String) throws -> String {
-        return signature
-    }
-
-    var signature: String!
-}
